+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # UNRELEASED
 
--   Adds a new API `Device#encryptToDeviceEvent` to encrypt an event using
-    olm.
+-   Adds a new API `Device#encryptToDeviceEvent` to encrypt a to-device event using
+    Olm.
 
 # matrix-sdk-crypto-wasm v4.6.0